+++ src/main/java/desafioTec/digix/dto/FamiliaRequestDto.java
@@ -2,5 +2,5 @@ package desafioTec.digix.dto;
 
 import java.util.List;
 
-public record FamiliaRequestDto(RepresentanteDto representante, int rendaTotal, List<DependenteDto> dependentes, ConjugeDto conjuge) {
+public record FamiliaRequestDto(RepresentanteDto representante, double rendaTotal, List<DependenteDto> dependentes, ConjugeDto conjuge) {
 }
