+++ src/main/java/desafioTec/digix/dto/FamiliaResponseDto.java
@@ -2,5 +2,5 @@ package desafioTec.digix.dto;
 
 import java.util.List;
 
-public record FamiliaResponseDto(RepresentanteDto representante, int rendaTotal, List<DependenteDto> dependentes, ConjugeDto conjuge) {
+public record FamiliaResponseDto(RepresentanteDto representante, double rendaTotal, List<DependenteDto> dependentes, ConjugeDto conjuge) {
 }
